requirements.txt
black==23.3.0
django==4.2
djangorestframework==3.14.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.2.2
social-auth-app-django==5.2.0
djoser==2.2.0
pytest==7.3.1
pytest-cov
pylint==2.17.3
